app/requirements.txt
ccxt
pandas
ta
pyyaml
python-telegram-bot
